pyproject.toml
[project]
name = "quant-open"
version = "0.1.0"
description = "Open quant infra with closed-source strategy plugin interface"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "pyarrow>=14",
  "akshare>=1.12",
  "pyyaml>=6",
  "loguru>=0.7",
  "backtrader>=1.9",
  "tqdm>=4.66",
  "typer>=0.12",
]

[project.scripts]
quantopen = "quantopen.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
